Moved some test codes (mostly Tuple related) (#1569)
* Two remaining test classes in redis.clients.jedis package are moved to redis.clients.jedis.tests package.

* Used import static for assertEquals.
Removed unused imports.

* Moved a Tuple test method from SortedSetCommandsTest class.

* Removed redundant tests (already in tupleCompare).
